+++ glean/src/index/qt.ts
@@ -17,6 +17,7 @@ import EventMetricType from "../core/metrics/types/event.js";
 import LabeledMetricType from "../core/metrics/types/labeled.js";
 import QuantityMetricType from "../core/metrics/types/quantity.js";
 import StringMetricType from "../core/metrics/types/string.js";
+import StringListMetricType from "../core/metrics/types/string_list.js";
 import TextMetricType from "../core/metrics/types/text.js";
 import TimespanMetricType from "../core/metrics/types/timespan.js";
 import UUIDMetricType from "../core/metrics/types/uuid.js";
@@ -124,6 +125,7 @@ export default {
     LabeledMetricType,
     QuantityMetricType,
     StringMetricType,
+    StringListMetricType,
     TimespanMetricType,
     TextMetricType,
     UUIDMetricType,